+++ packages/terra-form-select/src/terra-dev-site/doc/form-select/SearchSelectField.12.doc.jsx
@@ -2,12 +2,12 @@
 import React from 'react';
 import DocTemplate from 'terra-doc-template';
 import ReadMe from '../../../../docs/SearchSelectField.md';
+
 import { name, version } from '../../../../package.json';
 
-// Component Source
-import SearchSelectFieldSrc from '!raw-loader!../../../../src/SearchSelectField';
-
 // Example Files
+import SearchSelectFieldPropsTable from '!terra-props-table-loader!../../../../src/SearchSelectField';
+import SearchSelectFieldSrc from '!raw-loader!../../../../src/SearchSelectField';
 import SelectFieldExample from '../example/search/SearchSelectField';
 import SelectFieldExampleSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/search/SearchSelectField';
 import RequiredSearchSelectFieldExample from '../example/search/RequiredSearchSelectField';
@@ -19,10 +19,13 @@ const DocPage = () => (
     packageVersion={version}
     readme={ReadMe}
     srcPath={`https://github.com/cerner/terra-core/tree/master/packages/${name}`}
-    propsTables={[{
-      componentName: 'Search Select Field',
-      componentSrc: SearchSelectFieldSrc,
-    }]}
+    propsTables={[
+      {
+        componentName: 'Search Select Field',
+        componentSrc: SearchSelectFieldSrc,
+        componentProps: SearchSelectFieldPropsTable,
+      },
+    ]}
     examples={[{
       title: 'Search Select Field Example',
       example: <SelectFieldExample />,