+++ packages/terra-responsive-element/src/terra-dev-site/doc/responsive-element/ResponsiveElement.1.doc.jsx
@@ -1,8 +1,9 @@
 import React from 'react';
 import DocTemplate from 'terra-doc-template';
 import ReadMe from '../../../../docs/README.md';
+import { name, version } from '../../../../package.json';
 
-/* eslint-disable import/no-webpack-loader-syntax, import/first, import/no-duplicates */
+import ResponsiveElementPropsTable from '!terra-props-table-loader!../../../../src/ResponsiveElement';
 import ResponsiveElementSrc from '!raw-loader!../../../../src/ResponsiveElement';
 import BreakpointExample from '../example/BreakpointExample';
 import BreakpointExampleSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/BreakpointExample';
@@ -13,13 +14,15 @@ import UncontrolledBreakpointExampleSrc from '!raw-loader!../../../../src/terra-
 
 const DocPage = () => (
   <DocTemplate
-    packageName="terra-responsive-element"
+    packageName={name}
+    packageVersion={version}
     readme={ReadMe}
     srcPath="https://github.com/cerner/terra-core/tree/master/packages/terra-responsive-element"
     propsTables={[
       {
         componentName: 'Responsive Element',
         componentSrc: ResponsiveElementSrc,
+        componentProps: ResponsiveElementPropsTable,
       },
     ]}
     examples={[