+++ packages/terra-overlay/src/terra-dev-site/doc/overlay/Overlay.1.doc.jsx
@@ -3,12 +3,13 @@ import DocTemplate from 'terra-doc-template';
 import ReadMe from '../../../../docs/README.md';
 import { name, version } from '../../../../package.json';
 
-// Component Source
-import OverlayContainerSrc from '!raw-loader!../../../../src/OverlayContainer';
-import OverlaySrc from '!raw-loader!../../../../src/Overlay';
-import LoadingOverlaySrc from '!raw-loader!../../../../src/LoadingOverlay';
-
 // Example Files
+import LoadingOverlayPropsTable from '!terra-props-table-loader!../../../../src/LoadingOverlay';
+import LoadingOverlaySrc from '!raw-loader!../../../../src/LoadingOverlay';
+import OverlayPropsTable from '!terra-props-table-loader!../../../../src/Overlay';
+import OverlaySrc from '!raw-loader!../../../../src/Overlay';
+import OverlayContainerPropsTable from '!terra-props-table-loader!../../../../src/OverlayContainer';
+import OverlayContainerSrc from '!raw-loader!../../../../src/OverlayContainer';
 import OverlayExample from '../example/OverlayExample';
 import OverlayExampleSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/OverlayExample';
 import LoadingOverlayExample from '../example/LoadingOverlayExample';
@@ -34,16 +35,19 @@ const DocPage = () => (
     ]}
     propsTables={[
       {
-        componentName: 'Overlay Container',
-        componentSrc: OverlayContainerSrc,
-      },
-      {
         componentName: 'Overlay',
         componentSrc: OverlaySrc,
+        componentProps: OverlayPropsTable,
+      },
+      {
+        componentName: 'Overlay Container',
+        componentSrc: OverlayContainerSrc,
+        componentProps: OverlayContainerPropsTable,
       },
       {
         componentName: 'Loading Overlay',
         componentSrc: LoadingOverlaySrc,
+        componentProps: LoadingOverlayPropsTable,
       },
     ]}
   />