+++ packages/terra-icon/CHANGELOG.md
@@ -1,4 +1,4 @@
-Changelog
+ChangeLog
 =========
 
 Unreleased
@@ -7,6 +7,7 @@ Unreleased
 3.19.0 - (September 3, 2019)
 ------------------
 ### Added
+* Props table markdown file import added to doc site page
 * Added ThemeIconTemplate to support tooltip for theme icons
 * Added Translations for IconConsultInstructionsForUse tooltip text