+++ packages/terra-icon/CHANGELOG.md
@@ -1,8 +1,23 @@
-Changelog
+ChangeLog
 =========
 
 Unreleased
 ----------
+
+3.20.0 - (September 6, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
+3.19.0 - (September 3, 2019)
+------------------
+### Added
+* Props table markdown file import added to doc site page
+* Added ThemeIconTemplate to support tooltip for theme icons
+* Added Translations for IconConsultInstructionsForUse tooltip text
+
+### Changed
+* Import for package version from package.json added to doc page
 
 3.18.0 - (August 21, 2019)
 ------------------